Select, Update ja All omissa luokissaan, delete ja update luodaan tui:ssa komentoina ja erikseen omana luokkanaan selectin sisällä
