+++ src/main/java/projekti/UI/TUI.java
@@ -1,32 +1,23 @@
 package projekti.UI;
 
-import projekti.UI.commands.Command;
-import projekti.UI.commands.CreateRecommendation;
-import projekti.UI.commands.DBHelper;
-import projekti.UI.commands.DeleteRecommendation;
-import projekti.UI.commands.RecHelper;
+import projekti.UI.commands.*;
 import projekti.db.Dao;
 import projekti.domain.Blog;
 import projekti.domain.Book;
-import projekti.domain.Book.Properties;
 import projekti.domain.Other;
-import projekti.domain.Recommendation;
 
 import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.function.Function;
-import projekti.domain.Property;
-import projekti.domain.RecommendationFactory;
 
 public class TUI {
     private IO io;
    
-
     private RecHelper rh;
     private DBHelper db;
     private Command create;
-    private DeleteRecommendation delete;
+    private Command update;
+    private Command delete;
+    private Command select;
+    private Command all;
 
     public TUI(
             Dao<Book, Integer> bookDAO,
@@ -39,25 +30,13 @@ public class TUI {
 
         this.create = new CreateRecommendation(rh, db);
         this.delete = new DeleteRecommendation(rh, db);
+        this.update = new UpdateRecommendation(rh, db);
+        this.select = new SelectRecommendation(rh, db);
+        this.all = new ListAll(rh, db);
+
         this.io = io;
         rh.updateIDList();
     }
-
-   
-    private boolean update(Recommendation recommendation) throws SQLException {
-        switch (recommendation.getType()) {
-            case "BOOK":
-                return db.getBookDAO().update((Book) recommendation);
-            case "BLOG":
-                return db.getBlogDAO().update((Blog) recommendation);
-            case "OTHER":
-                return db.getOtherDAO().update((Other) recommendation);
-            default:
-                throw new IllegalArgumentException("No retrieve definition for recommendation of type: " + recommendation.getType());
-        }
-    }
-
-   
 
     public void run() throws SQLException {
         io.println("Welcome to the reading recommendation app!");
@@ -87,17 +66,16 @@ public class TUI {
                 create.execute();
                 break;
             case "all":
-                listRecommendations();
+                all.execute();
                 break;
             case "end":
                 io.println("\nshutting down program");
                 break;
             case "select":
-                selectRecommendation();
+                select.execute();
                 break;
             case "update":
-                updateRecommendation(rh.askForRecommendation());
-                
+                update.execute();              
                 break;
             case "delete":
                 delete.execute();
@@ -105,106 +83,6 @@ public class TUI {
             default:
                 io.println("\nnon-supported command");
                 break;
-        }
-    }
-
-    private void selectRecommendation() throws SQLException {
-        Recommendation recommendation = rh.askForRecommendation();
-
-        String input = "";
-
-        selectionLoop:
-        while (!input.equals("return")) {
-            if (recommendation == null) {
-                return;
-            }
-            io.println(rh.getListID(recommendation) + recommendation.toStringWithDescription());
-            io.println();
-
-            io.println("\ncommands for the selected recommendation: ");
-            io.println("\tedit \tedit the recommendation");
-            io.println("\tdelete \tremove the recommendation");
-            io.println("\treturn \treturn to the main program");
-
-            input = io.getInput();
-            switch (input) {
-                case "edit":
-                    recommendation = updateRecommendation(recommendation);
-                    break;
-                case "delete":
-                    delete.execute(recommendation);
-                    break selectionLoop;
-                case "return":
-                    io.println();
-                    break;
-                default:
-                    io.println("\nunsupported command");
-                    break;
-            }
-        }
-    }
-
-    
-
-    private void listRecommendations() throws SQLException {
-        List<Recommendation> recommendations = getAllRecommendations();
-        rh.updateIDList(); 
-        for (int i = 0; i < recommendations.size(); i++) {
-            io.println(i + recommendations.get(i).toString());
-        }
-    }
-
-    private List<Recommendation> getAllRecommendations() throws SQLException {
-        List<Recommendation> recommendations = new ArrayList<>();
-        recommendations.addAll(db.getBookDAO().findAll());
-        recommendations.addAll(db.getBlogDAO().findAll());
-        recommendations.addAll(db.getOtherDAO().findAll());
-        return recommendations;
-    }
-
-    private Recommendation updateRecommendation(Recommendation recommendation) throws SQLException {
-        Function<Property, String> requestProperty = (Property property) -> {
-            io.print("enter new " + property.getName() + " (or empty input to leave it unchanged): ");
-            String userInput = io.getInput().trim();
-            if (userInput.isEmpty()) {
-
-                return (String) recommendation.getProperty(property).orElse("");
-            }
-
-            return userInput;
-        };
-
-        String recommendationType = recommendation.getType().toLowerCase();
-        Recommendation updatedRecommendation;
-        try {
-            updatedRecommendation = RecommendationFactory.create()
-                    .selectType(recommendationType)
-                    .whileMissingProperties(requestProperty)
-                    .build();
-        } catch (IllegalArgumentException ex) {
-            io.println("\n " + recommendationType + " recommendation was not updated.");
-            throw ex;
-        }
-
-        Integer ID = recommendation.getProperty(Properties.ID).orElse(null);
-        updatedRecommendation.addProperty(Properties.ID, ID);
-        ID = rh.getIDList().indexOf(ID);
-
-        if (rh.confirm("are you sure you want to update recommendation " + ID + "?")) {
-            if (update(updatedRecommendation)) {
-                io.println();
-                io.println("update successful");
-                rh.updateIDList();
-                return updatedRecommendation;
-            } else {
-                io.println();
-                io.println("update failed");
-                return recommendation;
-            }
-        } else {
-            io.println();
-            io.println("recommendation update canceled");
-            return recommendation;
         }
     }
 }
